requirements.txt
customtkinter
jdatetime
Pillow
